Clarify resources re worker windows
This actually resolves #1021 in the old filecoin-docs (https://github.com/filecoin-project/filecoin-docs/issues/1021). The issue title now refers to clarifying what happens in worker windows, but the original request was just to make this part of the docs make sense as a sentence, and the table below gives enough detail for the reader to calculate what could be done with a given set of resources, which was the other part of the question in original issue #1021.