Storage provider updates and copy edits (#49)
* Remove out of date info on control addresses

Referenced feature is in current release.

* Clarify resources re worker windows

This actually resolves #1021 in the old filecoin-docs (https://github.com/filecoin-project/filecoin-docs/issues/1021). The issue title now refers to clarifying what happens in worker windows, but the original request was just to make this part of the docs make sense as a sentence, and the table below gives enough detail for the reader to calculate what could be done with a given set of resources, which was the other part of the question in original issue #1021.

* Updated Lotus worker co-location section

Modified co-location info for multi GPU setups since support was added to last RC and appears to be moving forward. Also added use case (which I can confirm is feasible) for co-location related to parallel GPU task processing.

* Reorganized storage groups section

I found the new documentation on storage groups a bit confusing so I thought I'd attempt to help clarify it a bit. Also corrected some errors/typos in config file and storage attach command syntax. One thing that's still not clear (to me) from this documentation is what the use case would be for having more than one value in Groups for any given sectorstore file. It also might be helpful to suggest how storage groups could work in an environment where some storage locations are shared and others are local, as well as considerations when workers have multiple (local) sector stores, but I haven't spent enough time configuring my own storage groups to be able to add that at this point.

I moved the storage groups documentation under a new heading at the bottom for multi-worker setups, in hopes that eventually there might be more robust official documentation on best practices for multiple workers since that's what most people seem to doing.

* Adds minor formatting tweaks.

* Addresses broken relref links.

Co-authored-by: Johnny <9611008+johnnymatthews@users.noreply.github.com>