Adjust weights across Lotus-pages
Adjusts the weights across all the pages in /Lotus/....

Every sections add +100 in weight, while each page in a given section add +5 to leave room for additional pages in between. The first page of a section starts at x05.

This should allow the page link/button at the bottom to link to the next page in a top-to-bottom approach. The rebalancing of weight in this commit should leave the current structure intact.